requirements.txt
bcrypt==4.3.0
black==25.1.0
click==8.1.8
colorama==0.4.6
greenlet==3.1.1
mariadb==1.1.12
mypy-extensions==1.0.0
mysqlclient==2.2.7
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
python-dotenv==1.1.0
SQLAlchemy==2.0.39
termcolor==2.5.0
typing_extensions==4.12.2
